Latest Changes

+++ src/app/admin/products/componets/DeleteVariation.jsx
@@ -20,46 +20,47 @@ const DeleteVariation = ({
 }) => {
   const [loading, setLoading] = useState(false);
   const [open, setOpen] = useState(false);
-  const handleOpen = () =>
-    isVariation
-      ? setOpen(!open)
-      : setVariations(variations.filter((_, i) => i !== index));
+  const handleOpen = () => setOpen(!open);
   const router = useRouter();
 
   const del = async () => {
     setLoading(true);
-    try {
-      await fetch(`/api/variation/${variation_id}`, {
-        method: "DELETE",
-      }).then((resp) => {
-        if (resp.status === 201) {
-          enqueueSnackbar("Variation deleted successfully", {
-            variant: "success",
-            preventDuplicates: true,
-            anchorOrigin: {
-              vertical: "top",
-              horizontal: "right",
-            },
-            autoHideDuration: 3000,
-            style: {
-              background: "white",
-              color: "black",
-              borderRadius: ".5rem",
-              boxShadow:
-                "0 1px 3px 0 rgba(0, 0, 0, 0.1), 0 1px 2px 0 rgba(0, 0, 0, 0.06)",
-              padding: "0 4px",
-            },
-          });
-          setVariations(
-            variations.filter(
-              (variation) => variation.variation_id !== variation_id
-            )
-          );
-        }
-        router.refresh();
-      });
-    } catch (e) {
-      console.error(e);
+    if (isVariation) {
+      try {
+        await fetch(`/api/variation/${variation_id}`, {
+          method: "DELETE",
+        }).then((resp) => {
+          if (resp.status === 201) {
+            enqueueSnackbar("Variation deleted successfully", {
+              variant: "success",
+              preventDuplicates: true,
+              anchorOrigin: {
+                vertical: "top",
+                horizontal: "right",
+              },
+              autoHideDuration: 3000,
+              style: {
+                background: "white",
+                color: "black",
+                borderRadius: ".5rem",
+                boxShadow:
+                  "0 1px 3px 0 rgba(0, 0, 0, 0.1), 0 1px 2px 0 rgba(0, 0, 0, 0.06)",
+                padding: "0 4px",
+              },
+            });
+            setVariations(
+              variations.filter(
+                (variation) => variation.variation_id !== variation_id
+              )
+            );
+          }
+          router.refresh();
+        });
+      } catch (e) {
+        console.error(e);
+      }
+    } else {
+      setVariations(variations.filter((_, i) => i !== index));
     }
     setOpen(false);
     setLoading(false);
